fix(Layout):コンポーネント化対応
